backend_service/himaliyan_sentinel/requirements.txt
djangorestframework
shapely
pytz